fixed compiler warnings - initialization fo certain static obejcts need to be prepended by CompuCell3d:: namespace operator, replaced BasicException with CC3DException, added explicit cast to size_t for dimentions (from int objects)

+++ CompuCell3D/core/CompuCell3D/steppables/PDESolvers/OpenCL/NonlinearSolver.cpp
@@ -155,7 +155,7 @@ viennacl::vector<float> const &NonlinearSolver::minusGoalFunc(viennacl::vector<f
 	}
 
 	//size_t glob_size[]={f};
-	size_t glob_size[]= {dim.x, dim.y, dim.z};
+	size_t glob_size[]= {(size_t)dim.x, (size_t)dim.y, (size_t)dim.z};
 	cl_int err=getOCLHelper().EnqueueNDRangeKernel(m_minusG->getKernel(), 3, glob_size, NULL); 
 	if(err!=CL_SUCCESS){
 		std::stringstream sstr;
@@ -205,7 +205,7 @@ viennacl::vector<float> const &NonlinearSolver::prod(viennacl::vector<float> con
 	}
 
 	//size_t glob_size[]={f};
-	size_t glob_size[]= {dim.x, dim.y, dim.z};
+	size_t glob_size[]= {(size_t)dim.x, (size_t)dim.y, (size_t)dim.z};
 	cl_int err=getOCLHelper().EnqueueNDRangeKernel(m_Jacobian->getKernel(), 3, glob_size, NULL); 
 	if(err!=CL_SUCCESS){
 		std::stringstream sstr;
